Moved the companyUrlObject interface to its own file, specificly for interfaces.
