+++ cypress/interfaces/link_checker_interfaces.ts
@@ -1,0 +1,5 @@
+
+export interface companyUrlObject {
+    url :string,
+    companyName :string
+}
